C++Builder demos for 2009 and 2010. Three projects of the 2010 group fail to build, however when I open their .dfm they build fine :(
git-svn-id: svn://svn.overbyte.be/ics/branches/icsv7@458 17a045a4-f01d-714f-988e-c09671ad6f7c

+++ cpp/Internet/OverbyteIcsPingTst1.cpp
@@ -104,6 +104,6 @@ void __fastcall TPingTstForm::Ping1EchoReply(TObject *Sender, TObject *Icmp, int
     else
         DisplayMemo->Lines->Add("Received " + IntToStr(Ping1->Reply.DataSize) +
                               " bytes from " + Ping1->HostIP +
-                              " in " + IntToStr(Ping1->Reply.RTT) + " msecs");
+                              " in " + IntToStr((int)Ping1->Reply.RTT) + " msecs");
 }
 //---------------------------------------------------------------------------
